Add placeholder PG association search handler
